optimize: 1. move apiserver auth filter to filter file. 2. add no permission handle when adaptor permissions.
